+++ infra/Services/ComposeMessagesForProcessing.cs
@@ -41,8 +41,8 @@ namespace infra.Services
             msg = title + " " + candidateName + "<br>Application No.:" + candidate.ApplicationNo + ", PP No.:" + candidate.PpNo;
             msg += "email: " + candidate.Email + "<br><br>Dear Sir:<br><br>please be advised of the following updates to your departure formalities<br><br>";
             msg += "<tab>Date: " + deploy.TransactionDate.Date;
-            msg += "<br><tab>Transaction: " + _commonServices.DeploymentStageNameFromStageId(deploy.StageId);
-            msg += deploy.NextStageId > 0 ? "<br><tab>Next Stage:" + _commonServices.DeploymentStageNameFromStageId(deploy.NextStageId) + ", scheduled by " + deploy.NextEstimatedStageDate.Date : "";
+            msg += "<br><tab>Transaction: " + _commonServices.DeploymentStageNameFromStageId((int)deploy.StageId);
+            msg += deploy.NextStageId > 0 ? "<br><tab>Next Stage:" + _commonServices.DeploymentStageNameFromStageId((int)deploy.NextStageId) + ", scheduled by " + deploy.NextEstimatedStageDate.Date : "";
             msg += "<br><br>Regards/Processing Divn";
 
             var emailMessage = new EmailMessage
@@ -80,8 +80,8 @@ namespace infra.Services
             msg = title + " " + candidateName + ", PP No.:" + candidate.PpNo;
             msg += "pl be advised of flg updates to yr departure formalities<br><br>";
             msg += "<tab>Date: " + deploy.TransactionDate.Date;
-            msg += "<br><tab>Transaction: " + _commonServices.DeploymentStageNameFromStageId(deploy.StageId);
-            msg += deploy.NextStageId > 0 ? "<br><tab>Next Stage:" + _commonServices.DeploymentStageNameFromStageId(deploy.NextStageId) + ", scheduled by " + deploy.NextEstimatedStageDate.Date : "";
+            msg += "<br><tab>Transaction: " + _commonServices.DeploymentStageNameFromStageId((int)deploy.StageId);
+            msg += deploy.NextStageId > 0 ? "<br><tab>Next Stage:" + _commonServices.DeploymentStageNameFromStageId((int)deploy.NextStageId) + ", scheduled by " + deploy.NextEstimatedStageDate.Date : "";
             msg += "<br><br>Regards/Processing Divn";
 
             var smsMessage = new SMSMessage
